app/loadtest/requirements.txt
locustio
uuid==1.30
